move TorServicePrefs class to topl-service-base module
