+++ topl-service-base/src/main/java/io/matthewnelson/topl_service_base/TorServicePrefs.kt
@@ -1,87 +1,19 @@
-/*
-* TorOnionProxyLibrary-Android (a.k.a. topl-android) is a derivation of
-* work from the Tor_Onion_Proxy_Library project that started at commit
-* hash `74407114cbfa8ea6f2ac51417dda8be98d8aba86`. Contributions made after
-* said commit hash are:
-*
-*     Copyright (C) 2020 Matthew Nelson
-*
-*     This program is free software: you can redistribute it and/or modify it
-*     under the terms of the GNU General Public License as published by the
-*     Free Software Foundation, either version 3 of the License, or (at your
-*     option) any later version.
-*
-*     This program is distributed in the hope that it will be useful, but
-*     WITHOUT ANY WARRANTY; without even the implied warranty of MERCHANTABILITY
-*     or FITNESS FOR A PARTICULAR PURPOSE. See the GNU General Public License
-*     for more details.
-*
-*     You should have received a copy of the GNU General Public License
-*     along with this program. If not, see <https://www.gnu.org/licenses/>.
-*
-* `===========================================================================`
-* `+++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++`
-* `===========================================================================`
-*
-* The following exception is an additional permission under section 7 of the
-* GNU General Public License, version 3 (“GPLv3”).
-*
-*     "The Interfaces" is henceforth defined as Application Programming Interfaces
-*     that are publicly available classes/functions/etc (ie: do not contain the
-*     visibility modifiers `internal`, `private`, `protected`, or are within
-*     classes/functions/etc that contain the aforementioned visibility modifiers)
-*     to TorOnionProxyLibrary-Android users that are needed to implement
-*     TorOnionProxyLibrary-Android and reside in ONLY the following modules:
-*
-*      - topl-core-base
-*      - topl-service
-*
-*     The following are excluded from "The Interfaces":
-*
-*       - All other code
-*
-*     Linking TorOnionProxyLibrary-Android statically or dynamically with other
-*     modules is making a combined work based on TorOnionProxyLibrary-Android.
-*     Thus, the terms and conditions of the GNU General Public License cover the
-*     whole combination.
-*
-*     As a special exception, the copyright holder of TorOnionProxyLibrary-Android
-*     gives you permission to combine TorOnionProxyLibrary-Android program with free
-*     software programs or libraries that are released under the GNU LGPL and with
-*     independent modules that communicate with TorOnionProxyLibrary-Android solely
-*     through "The Interfaces". You may copy and distribute such a system following
-*     the terms of the GNU GPL for TorOnionProxyLibrary-Android and the licenses of
-*     the other code concerned, provided that you include the source code of that
-*     other code when and as the GNU GPL requires distribution of source code and
-*     provided that you do not modify "The Interfaces".
-*
-*     Note that people who make modified versions of TorOnionProxyLibrary-Android
-*     are not obligated to grant this special exception for their modified versions;
-*     it is their choice whether to do so. The GNU General Public License gives
-*     permission to release a modified version without this exception; this exception
-*     also makes it possible to release a modified version which carries forward this
-*     exception. If you modify "The Interfaces", this exception does not apply to your
-*     modified version of TorOnionProxyLibrary-Android, and you must remove this
-*     exception when you distribute your modified version.
-* */
-package io.matthewnelson.topl_service.prefs
+package io.matthewnelson.topl_service_base
 
 import android.content.Context
 import android.content.SharedPreferences
 import androidx.annotation.WorkerThread
-import io.matthewnelson.topl_service.util.ServiceConsts
-import io.matthewnelson.topl_service_base.BaseServiceConsts
 
 /**
  * This class provides a standardized way for library users to change settings used
- * by [io.matthewnelson.topl_service.service.TorService] such that the values expressed
- * as default [io.matthewnelson.topl_core_base.TorSettings] when initializing things via
- * the [io.matthewnelson.topl_service.TorServiceController.Builder] can be updated. The
- * values saved to [TorServicePrefs] are always preferred over the defaults declared.
+ * by the `topl-service` module such that the values expressed as default
+ * [io.matthewnelson.topl_core_base.TorSettings] when initializing things via can be modified
+ * by the implementing application.
  *
- * See [io.matthewnelson.topl_service.service.components.onionproxy.ServiceTorSettings]
+ * The values saved to [TorServicePrefs] are always preferred over the defaults declared
+ * when initializing the `topl-service` module.
  * */
-class TorServicePrefs(context: Context): ServiceConsts() {
+class TorServicePrefs(context: Context): BaseServiceConsts() {
 
     companion object {
         const val TOR_SERVICE_PREFS_NAME = "TorServicePrefs"
@@ -89,8 +21,9 @@ class TorServicePrefs(context: Context): ServiceConsts() {
         const val NULL_STRING_VALUE = "NULL_STRING_VALUE"
     }
 
-    private val prefs: SharedPreferences =
+    private val prefs: SharedPreferences by lazy {
         context.getSharedPreferences(TOR_SERVICE_PREFS_NAME, Context.MODE_PRIVATE)
+    }
 
 
     /////////////////
